Effects.Concurrent.STM

Description

Provides classes for STM.

Since: 0.1

Effect

data STM a

A monad supporting atomic memory transactions.

class Monad m => MonadSTM (m :: Type -> Type) where

STM effect. Note that this class is for monads that can lift entire STM transactions (i.e. atomically). It is not intended for "STM-like" monads -- hence has no STM instance -- as the semantics for "STM-like" and "can lift STM atomically" are different.

Since: 0.1

Methods

atomically :: HasCallStack => STM a -> m a

Lifted atomically.

Since: 0.1

TVar

data TVar a

Shared memory locations that support atomic memory transactions.

newTVarA :: (HasCallStack, MonadSTM m) => a -> m (TVar a)

Create a new TVar holding a value supplied and lifts the result via atomically.

Since: 0.1

readTVarA :: (HasCallStack, MonadSTM m) => TVar a -> m a

Return the current value stored in a TVar and lifts the result via atomically.

Since: 0.1

writeTVarA :: (HasCallStack, MonadSTM m) => TVar a -> a -> m ()

Write the supplied value into a TVar and lifts the action via atomically.

Since: 0.1

modifyTVarA' :: (HasCallStack, MonadSTM m) => TVar a -> (a -> a) -> m ()

Strict version of modifyTVar, lifting the action via atomically.

Since: 0.1

TBQueue

data TBQueue a

TBQueue is an abstract type representing a bounded FIFO channel.

Since: stm-2.4

newTBQueueA

Arguments

:: (HasCallStack, MonadSTM m)

=> Natural
maximum number of elements the queue can hold

-> m (TBQueue a)

Builds and returns a new instance of TBQueue, lifting via atomically.

Since: 0.1

readTBQueueA :: (HasCallStack, MonadSTM m) => TBQueue a -> m a

Read the next value from the TBQueue, lifting via atomically.

Since: 0.1

tryReadTBQueueA :: (HasCallStack, MonadSTM m) => TBQueue a -> m (Maybe a)

A version of readTBQueue which does not retry. Instead it returns Nothing if no value is available. Lifts via atomically.

Since: 0.1

writeTBQueueA :: (HasCallStack, MonadSTM m) => TBQueue a -> a -> m ()

Write a value to a TBQueue; blocks if the queue is full. Lifts via atomically.

Since: 0.1

flushTBQueueA :: (HasCallStack, MonadSTM m) => TBQueue a -> m [a]

Efficiently read the entire contents of a TBQueue into a list. This function never retries. Lifts via atomically.

Since: 0.1

Reexports

data Natural

Natural number

Invariant: numbers <= 0xffffffffffffffff use the NS constructor