Effects.Concurrent.STM

Description

Provides classes for STM.

Since: 0.1

Effect

data STM a

A monad supporting atomic memory transactions.

class Monad m => MonadSTM m where

STM effect.

Since: 0.1

Methods

atomically :: HasCallStack => STM a -> m a

Perform a series of STM actions atomically.

Using atomically inside an unsafePerformIO or unsafeInterleaveIO subverts some of guarantees that STM provides. It makes it possible to run a transaction inside of another transaction, depending on when the thunk is evaluated. If a nested transaction is attempted, an exception is thrown by the runtime. It is possible to safely use atomically inside unsafePerformIO or unsafeInterleaveIO, but the typechecker does not rule out programs that may attempt nested transactions, meaning that the programmer must take special care to prevent these.

However, there are functions for creating transactional variables that can always be safely called in unsafePerformIO. See: newTVarIO, newTChanIO, newBroadcastTChanIO, newTQueueIO, newTBQueueIO, and newTMVarIO.

Using unsafePerformIO inside of atomically is also dangerous but for different reasons. See unsafeIOToSTM for more on this.

Since: 0.1

TVar

data TVar a

Shared memory locations that support atomic memory transactions.

newTVarA :: (HasCallStack, MonadSTM m) => a -> m (TVar a)

Create a new TVar holding a value supplied and lifts the result via atomically.

Since: 0.1

readTVarA :: (HasCallStack, MonadSTM m) => TVar a -> m a

Return the current value stored in a TVar and lifts the result via atomically.

Since: 0.1

writeTVarA :: (HasCallStack, MonadSTM m) => TVar a -> a -> m ()

Write the supplied value into a TVar and lifts the action via atomically.

Since: 0.1

modifyTVarA' :: (HasCallStack, MonadSTM m) => TVar a -> (a -> a) -> m ()

Strict version of modifyTVar, lifting the action via atomically.

Since: 0.1

TBQueue

data TBQueue a

TBQueue is an abstract type representing a bounded FIFO channel.

Since: stm-2.4

newTBQueueA

Arguments

:: (HasCallStack, MonadSTM m)

=> Natural
maximum number of elements the queue can hold

-> m (TBQueue a)

Builds and returns a new instance of TBQueue, lifting via atomically.

Since: 0.1

readTBQueueA :: (HasCallStack, MonadSTM m) => TBQueue a -> m a

Read the next value from the TBQueue, lifting via atomically.

Since: 0.1

tryReadTBQueueA :: (HasCallStack, MonadSTM m) => TBQueue a -> m (Maybe a)

A version of readTBQueue which does not retry. Instead it returns Nothing if no value is available. Lifts via atomically.

Since: 0.1

writeTBQueueA :: (HasCallStack, MonadSTM m) => TBQueue a -> a -> m ()

Write a value to a TBQueue; blocks if the queue is full. Lifts via atomically.

Since: 0.1

flushTBQueueA :: (HasCallStack, MonadSTM m) => TBQueue a -> m [a]

Efficiently read the entire contents of a TBQueue into a list. This function never retries. Lifts via atomically.

Since: 0.1

Reexports

data Natural

Natural number

Invariant: numbers <= 0xffffffffffffffff use the NS constructor